Make RPC scaffolding available for Node
